chore: clean up publish findings
