+++ app/agent/stages/publish_findings/delivery/__init__.py
@@ -1,0 +1,5 @@
+"""Delivery dispatch for publish findings."""
+
+from app.agent.stages.publish_findings.delivery.dispatch import dispatch_report
+
+__all__ = ["dispatch_report"]
