+++ app/agent/stages/publish_findings/formatters/__init__.py
@@ -3,15 +3,23 @@
 from app.agent.stages.publish_findings.formatters.evidence import (
     format_cited_evidence_section,
 )
-from app.agent.stages.publish_findings.formatters.infrastructure import (
-    format_infrastructure_correlation,
+from app.agent.stages.publish_findings.formatters.messages import (
+    ReportMessages,
+    build_report_messages,
 )
-from app.agent.stages.publish_findings.formatters.lineage import format_data_lineage_flow
-from app.agent.stages.publish_findings.formatters.report import format_slack_message
+from app.agent.stages.publish_findings.formatters.report import (
+    build_slack_blocks,
+    format_slack_message,
+    format_telegram_message,
+    format_whatsapp_message,
+)
 
 __all__ = [
+    "ReportMessages",
+    "build_report_messages",
+    "build_slack_blocks",
     "format_slack_message",
+    "format_telegram_message",
+    "format_whatsapp_message",
     "format_cited_evidence_section",
-    "format_infrastructure_correlation",
-    "format_data_lineage_flow",
 ]
