+++ app/agent/stages/publish_findings/formatters/channels/__init__.py
@@ -1,0 +1,15 @@
+"""Channel-specific report formatters."""
+
+from app.agent.stages.publish_findings.formatters.channels.slack import (
+    build_slack_blocks,
+    format_slack_message,
+)
+from app.agent.stages.publish_findings.formatters.channels.telegram import format_telegram_message
+from app.agent.stages.publish_findings.formatters.channels.whatsapp import format_whatsapp_message
+
+__all__ = [
+    "build_slack_blocks",
+    "format_slack_message",
+    "format_telegram_message",
+    "format_whatsapp_message",
+]
