+++ app/agent/stages/publish_findings/formatters/channels/slack.py
@@ -1,0 +1,8 @@
+"""Slack report formatting."""
+
+from app.agent.stages.publish_findings.formatters.report import (
+    build_slack_blocks,
+    format_slack_message,
+)
+
+__all__ = ["build_slack_blocks", "format_slack_message"]
