+++ app/agent/stages/publish_findings/formatters/channels/telegram.py
@@ -1,0 +1,5 @@
+"""Telegram report formatting."""
+
+from app.agent.stages.publish_findings.formatters.report import format_telegram_message
+
+__all__ = ["format_telegram_message"]
