+++ app/agent/stages/publish_findings/formatters/channels/whatsapp.py
@@ -1,0 +1,5 @@
+"""WhatsApp and SMS report formatting."""
+
+from app.agent.stages.publish_findings.formatters.report import format_whatsapp_message
+
+__all__ = ["format_whatsapp_message"]
