+++ app/agent/stages/publish_findings/formatters/evidence.py
@@ -4,12 +4,12 @@ import html
 from collections.abc import Callable
 from typing import Any
 
+from app.agent.stages.publish_findings.context import ReportContext
 from app.agent.stages.publish_findings.formatters.base import (
     format_html_link,
     format_slack_link,
     shorten_text,
 )
-from app.agent.stages.publish_findings.report_context import ReportContext
 from app.agent.stages.publish_findings.urls.aws import (
     build_datadog_logs_url,
     build_grafana_explore_url,
