+++ app/agent/stages/publish_findings/node.py
@@ -1,302 +1,59 @@
 """Main orchestration node for report generation and publishing."""
 
-import logging
 from typing import Any
 
-from app.agent.stages.publish_findings.formatters.report import (
-    build_slack_blocks,
-    format_slack_message,
-    format_telegram_message,
-    format_whatsapp_message,
+from app.agent.stages.publish_findings.context import build_report_context
+from app.agent.stages.publish_findings.delivery import dispatch_report
+from app.agent.stages.publish_findings.formatters.messages import (
+    ReportMessages,
+    build_report_messages,
 )
-from app.agent.stages.publish_findings.gitlab_writeback import post_gitlab_mr_writeback
 from app.agent.stages.publish_findings.renderers.editor import open_in_editor
 from app.agent.stages.publish_findings.renderers.terminal import render_report
-from app.agent.stages.publish_findings.report_context import build_report_context
 from app.masking import MaskingContext
 from app.state import InvestigationState
-from app.types.config import NodeConfig
 from app.utils.ingest_delivery import create_investigation_and_attach_url
-from app.utils.tracing import traceable
-
-logger = logging.getLogger(__name__)
 
 
-def generate_report(state: InvestigationState) -> dict:
+def generate_report(
+    state: InvestigationState,
+    *,
+    render_terminal: bool = True,
+    open_editor: bool = True,
+) -> dict[str, Any]:
     """Generate and publish the final RCA report."""
-    from app.utils.slack_delivery import build_action_blocks, send_slack_report
-
     ctx = build_report_context(state)
     short_summary = state.get("problem_md")
-    slack_message = format_slack_message(ctx)
+    messages = build_report_messages(ctx)
 
     # Restore any masked infrastructure identifiers in user-facing output.
     # No-op when masking is disabled or the state has no placeholders.
     masking_ctx = MaskingContext.from_state(dict(state))
-    slack_message = masking_ctx.unmask(slack_message)
+    messages = ReportMessages(
+        slack_text=masking_ctx.unmask(messages.slack_text),
+        telegram_html=masking_ctx.unmask(messages.telegram_html),
+        whatsapp_text=masking_ctx.unmask(messages.whatsapp_text),
+        slack_blocks=masking_ctx.unmask_value(messages.slack_blocks),
+    )
     if isinstance(short_summary, str):
         short_summary = masking_ctx.unmask(short_summary)
 
     investigation_id, investigation_url = create_investigation_and_attach_url(
         state,
-        slack_message,
+        messages.slack_text,
         short_summary,
     )
 
-    telegram_message = masking_ctx.unmask(format_telegram_message(ctx))
-    whatsapp_message = masking_ctx.unmask(format_whatsapp_message(ctx))
+    if render_terminal:
+        render_report(messages.slack_text, root_cause_category=state.get("root_cause_category"))
+    if open_editor:
+        open_in_editor(messages.slack_text)
 
-    all_blocks = build_slack_blocks(ctx) + build_action_blocks(investigation_url, investigation_id)
-    all_blocks = masking_ctx.unmask_value(all_blocks)
-    render_report(slack_message, root_cause_category=state.get("root_cause_category"))
-    open_in_editor(slack_message)
-
-    slack_ctx = state.get("slack_context", {})
-    thread_ts = slack_ctx.get("thread_ts") or slack_ctx.get("ts")
-    _channel = slack_ctx.get("channel_id")
-    _token = slack_ctx.get("access_token")
-    _alert_ts = slack_ctx.get("ts") or slack_ctx.get("thread_ts")
-
-    resolved = state.get("resolved_integrations") or {}
-    discord_creds = resolved.get("discord", {})
-    logger.debug("[publish] slack_ctx=%s", slack_ctx)
-    logger.debug(
-        "[publish] discord creds present=%s keys=%s",
-        bool(discord_creds),
-        list(discord_creds.keys()) if discord_creds else [],
+    dispatch_report(
+        state,
+        messages,
+        investigation_id=investigation_id,
+        investigation_url=investigation_url,
     )
 
-    report_posted, delivery_error = send_slack_report(
-        slack_message,
-        channel=_channel,
-        thread_ts=thread_ts,
-        access_token=_token,
-        blocks=all_blocks,
-    )
-
-    logger.debug(
-        "[publish] slack delivery: posted=%s channel=%s thread_ts=%s error=%s",
-        report_posted,
-        _channel,
-        thread_ts,
-        delivery_error,
-    )
-    if report_posted and _token and _channel and _alert_ts:
-        from app.utils.slack_delivery import swap_reaction
-
-        swap_reaction("eyes", "clipboard", _channel, _alert_ts, _token)
-    elif thread_ts and not report_posted:
-        raise RuntimeError(
-            f"[publish] Slack delivery failed: channel={_channel}, thread_ts={thread_ts}, reason={delivery_error}"
-        )
-
-    # Discord delivery — uses integration credentials if configured
-    if discord_creds:
-        from app.utils.discord_delivery import send_discord_report
-
-        discord_ctx = state.get("discord_context") or {}
-        bot_token = discord_ctx.get("bot_token") or discord_creds.get("bot_token", "")
-        channel_id = discord_ctx.get("channel_id") or discord_creds.get("default_channel_id", "")
-        thread_id = discord_ctx.get("thread_id", "")
-        logger.debug(
-            "[publish] discord delivery: channel_id=%s thread_id=%s bot_token_present=%s",
-            channel_id,
-            thread_id,
-            bool(bot_token),
-        )
-        if bot_token and channel_id:
-            discord_posted, discord_error = send_discord_report(
-                slack_message,
-                {"bot_token": bot_token, "channel_id": channel_id, "thread_id": thread_id},
-            )
-            logger.debug(
-                "[publish] discord delivery: posted=%s error=%s", discord_posted, discord_error
-            )
-            if not discord_posted:
-                logger.warning(
-                    "[publish] Discord delivery failed: channel=%s error=%s",
-                    channel_id,
-                    discord_error,
-                )
-        else:
-            logger.debug(
-                "[publish] discord delivery: skipped — bot_token_present=%s channel_id=%s",
-                bool(bot_token),
-                channel_id,
-            )
-    else:
-        logger.debug("[publish] discord delivery: no discord integration configured")
-
-    # Telegram delivery — uses integration credentials if configured
-    telegram_creds = resolved.get("telegram", {})
-    if telegram_creds:
-        from app.utils.telegram_delivery import send_telegram_report
-
-        telegram_ctx = state.get("telegram_context") or {}
-        bot_token = telegram_ctx.get("bot_token") or telegram_creds.get("bot_token", "")
-        chat_id = telegram_ctx.get("chat_id") or telegram_creds.get("default_chat_id", "")
-        reply_to = str(telegram_ctx.get("reply_to_message_id") or "")
-        logger.debug(
-            "[publish] telegram delivery: chat_id=%s reply_to=%s bot_token_present=%s",
-            chat_id,
-            reply_to,
-            bool(bot_token),
-        )
-        if bot_token and chat_id:
-            tg_posted, tg_error = send_telegram_report(
-                telegram_message,
-                {"bot_token": bot_token, "chat_id": chat_id, "reply_to_message_id": reply_to},
-            )
-            logger.debug("[publish] telegram delivery: posted=%s error=%s", tg_posted, tg_error)
-            if not tg_posted:
-                logger.warning(
-                    "[publish] Telegram delivery failed: chat_id=%s error=%s",
-                    chat_id,
-                    tg_error,
-                )
-        else:
-            logger.debug(
-                "[publish] telegram delivery: skipped — bot_token_present=%s chat_id=%s",
-                bool(bot_token),
-                chat_id,
-            )
-    else:
-        logger.debug("[publish] telegram delivery: no telegram integration configured")
-
-    # WhatsApp delivery — uses integration credentials if configured
-    whatsapp_creds = resolved.get("whatsapp", {})
-    if whatsapp_creds:
-        from app.utils.whatsapp_delivery import send_whatsapp_report
-
-        _wa_ctx: dict[str, Any] = state.get("whatsapp_context") or {}
-        account_sid = _wa_ctx.get("account_sid") or whatsapp_creds.get("account_sid", "")
-        auth_token = _wa_ctx.get("auth_token") or whatsapp_creds.get("auth_token", "")
-        from_number = _wa_ctx.get("from_number") or whatsapp_creds.get("from_number", "")
-        to = _wa_ctx.get("to") or whatsapp_creds.get("default_to", "")
-        logger.debug(
-            "[publish] whatsapp delivery: to=%s account_sid=%s auth_token_present=%s from_number=%s",
-            to,
-            account_sid,
-            bool(auth_token),
-            from_number,
-        )
-        if account_sid and auth_token and from_number and to:
-            wa_posted, wa_error = send_whatsapp_report(
-                whatsapp_message,
-                {
-                    "account_sid": account_sid,
-                    "auth_token": auth_token,
-                    "from_number": from_number,
-                    "to": to,
-                },
-            )
-            logger.debug("[publish] whatsapp delivery: posted=%s error=%s", wa_posted, wa_error)
-            if not wa_posted:
-                logger.warning(
-                    "[publish] WhatsApp delivery failed: to=%s error=%s",
-                    to,
-                    wa_error,
-                )
-        else:
-            logger.debug(
-                "[publish] whatsapp delivery: skipped — account_sid_present=%s auth_token_present=%s from_number_present=%s to_present=%s",
-                bool(account_sid),
-                bool(auth_token),
-                bool(from_number),
-                bool(to),
-            )
-    else:
-        logger.debug("[publish] whatsapp delivery: no whatsapp integration configured")
-
-    # Twilio SMS — dispatched independently of the legacy WhatsApp record
-    # above. WhatsApp delivery is owned solely by the ``whatsapp`` integration.
-    twilio_creds = resolved.get("twilio", {})
-    if twilio_creds:
-        sms_cfg = twilio_creds.get("sms") or {}
-        if sms_cfg.get("enabled"):
-            from app.utils.twilio_delivery import send_twilio_sms_report
-
-            twilio_sms_ctx: dict[str, Any] = state.get("twilio_sms_context") or {}
-            sms_to = twilio_sms_ctx.get("to") or sms_cfg.get("default_to") or ""
-            sms_from = sms_cfg.get("from_number", "")
-            messaging_service_sid = sms_cfg.get("messaging_service_sid", "")
-            account_sid = twilio_creds.get("account_sid", "")
-            auth_token = twilio_creds.get("auth_token", "")
-            logger.debug(
-                "[publish] twilio sms delivery: to=%s from=%s msg_service=%s account_sid_present=%s",
-                sms_to,
-                sms_from,
-                messaging_service_sid,
-                bool(account_sid),
-            )
-            if account_sid and auth_token and sms_to and (sms_from or messaging_service_sid):
-                # SMS currently reuses the WhatsApp-formatted body — both
-                # channels render the same plain-text RCA summary. If
-                # formatting needs to diverge (e.g. shorter SMS body or
-                # different headers), introduce ``format_sms_message`` in
-                # ``formatters/report.py`` and route it through here.
-                sms_message = whatsapp_message
-                sms_ok, sms_error, sms_sid = send_twilio_sms_report(
-                    sms_message,
-                    {
-                        "account_sid": account_sid,
-                        "auth_token": auth_token,
-                        "from_number": sms_from,
-                        "messaging_service_sid": messaging_service_sid,
-                        "to": sms_to,
-                    },
-                )
-                logger.debug(
-                    "[publish] twilio sms delivery: posted=%s sid=%s error=%s",
-                    sms_ok,
-                    sms_sid,
-                    sms_error,
-                )
-                if not sms_ok:
-                    logger.warning(
-                        "[publish] Twilio SMS delivery failed: to=%s error=%s",
-                        sms_to,
-                        sms_error,
-                    )
-            else:
-                # SMS channel is enabled but something required for delivery
-                # is missing — most commonly no recipient (default_to unset and
-                # no runtime twilio_sms_context.to). Warn so the misconfig is
-                # visible rather than silently swallowed.
-                logger.warning(
-                    "[publish] twilio sms delivery: skipped — SMS channel is enabled "
-                    "but not deliverable (recipient_present=%s sender_present=%s "
-                    "account_sid_present=%s auth_token_present=%s). "
-                    "Set TWILIO_SMS_DEFAULT_TO to enable auto-delivery.",
-                    bool(sms_to),
-                    bool(sms_from or messaging_service_sid),
-                    bool(account_sid),
-                    bool(auth_token),
-                )
-    else:
-        logger.debug("[publish] twilio delivery: no twilio integration configured")
-
-    openclaw_creds = resolved.get("openclaw", {})
-    if openclaw_creds:
-        from app.utils.openclaw_delivery import send_openclaw_report
-
-        oc_posted, oc_error = send_openclaw_report(state, slack_message, openclaw_creds)
-        logger.debug("[publish] openclaw delivery: posted=%s error=%s", oc_posted, oc_error)
-        if not oc_posted:
-            logger.debug("[publish] OpenClaw delivery failed: %s", oc_error)
-    else:
-        logger.debug("[publish] openclaw delivery: no openclaw integration configured")
-
-    post_gitlab_mr_writeback(state, slack_message)
-
-    return {"slack_message": slack_message, "report": slack_message}
-
-
-@traceable(name="node_publish_findings")
-def node_publish_findings(
-    state: InvestigationState,
-    config: NodeConfig | None = None,
-) -> dict:
-    """Publish step wrapper with optional tracing."""
-    del config
-    return generate_report(state)
+    return {"slack_message": messages.slack_text, "report": messages.slack_text}
