+++ tests/delivery/test_report_provenance.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
+from app.agent.stages.publish_findings.context import build_report_context
 from app.agent.stages.publish_findings.formatters.report import (
     build_slack_blocks,
     format_slack_message,
     format_telegram_message,
 )
-from app.agent.stages.publish_findings.report_context import build_report_context
 
 
 def _make_state() -> dict:
